refactor: gig description header
