+++ app/(modals)/gig-description.tsx
@@ -1,21 +1,12 @@
-import {
-  Image,
-  ScrollView,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-} from "react-native";
+import { Image, ScrollView, StyleSheet, Text, View } from "react-native";
 import React from "react";
 import tw from "twrnc";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import { Feather, Ionicons } from "@expo/vector-icons";
 import PrimaryButton from "@/components/buttons/PrimaryButton";
-import { useNavigation } from "expo-router";
+import GigDescriptionHeader from "@/components/navigation/headers/GigDescriptionHeader";
 
 const GigDescription = () => {
   const insets = useSafeAreaInsets();
-  const navigation = useNavigation();
   return (
     <ScrollView
       contentContainerStyle={[
@@ -26,21 +17,7 @@ const GigDescription = () => {
       ]}
       style={tw`h-full w-full`}
     >
-      <View style={tw`flex flex-row gap-4 py-2`}>
-        <TouchableOpacity
-          onPress={() => navigation.goBack()}
-          style={tw`bg-zinc-100 rounded-full p-2`}
-        >
-          <Ionicons name="arrow-back-sharp" size={24} color="#71717a" />
-        </TouchableOpacity>
-        <View style={tw`flex-1`} />
-        <TouchableOpacity style={tw`bg-zinc-100 rounded-full p-2`}>
-          <Feather name="share-2" size={24} color="#71717a" />
-        </TouchableOpacity>
-        <TouchableOpacity style={tw`bg-zinc-100 rounded-full p-2`}>
-          <Feather name="bookmark" size={24} color="#71717a" />
-        </TouchableOpacity>
-      </View>
+      <GigDescriptionHeader />
       <View style={tw`text-bg-zinc-100 rounded-2xl h-32 w-32`}>
         <Image
           style={tw`h-32 w-32 bg-zinc-100 rounded-2xl`}
